requirements.txt
Flask>=1.1<2
google-cloud-bigquery>=1.24.0<1.3
Jinja2>=2.10<3
jinjasql>= 0.1.7<0.2
# pandas>0.25<0.3